requirements.txt
git+git://github.com/pavius/py-opensprinkler@support-logs
PyYAML==5.4.1
aiogram==2.14.3
uvloop==0.15.3
pandas==1.3.0
matplotlib==3.4.2
numpy==1.21.1
sendgrid==6.7.1
aiocron==1.6
